+++ kvm_control.py
@@ -1,0 +1,26 @@
+# kvm_control.py
+from pynput import mouse, keyboard
+
+class KVMControl:
+    def __init__(self):
+        self.shortcut = None
+        self.active = False
+
+    def set_shortcut(self, shortcut):
+        self.shortcut = shortcut
+
+    def toggle_control(self):
+        self.active = not self.active
+        # Implement control switching logic here
+
+    def start_listening(self):
+        # Start listening for keyboard and mouse events
+        pass
+
+    def send_event(self, event):
+        # Send event to the other machine
+        pass
+
+    def process_event(self, event):
+        # Process received event
+        pass
