+++ network.py
@@ -1,0 +1,37 @@
+# network.py
+import socket
+import netifaces
+
+class NetworkManager:
+    def __init__(self):
+        self.port = 0
+        self.broadcast_name = "default kvm"
+        self.interface = None
+
+    def get_interfaces(self):
+        return netifaces.interfaces()
+
+    def set_interface(self, interface):
+        self.interface = interface
+
+    def set_port(self, port):
+        self.port = int(port) if port else 0
+
+    def set_broadcast_name(self, name):
+        self.broadcast_name = name
+
+    def discover_hosts(self):
+        # Implement host discovery logic here
+        pass
+
+    def start_host(self):
+        # Implement host server logic here
+        pass
+
+    def connect_to_host(self, host):
+        # Implement client connection logic here
+        pass
+
+    def p2p_connect(self):
+        # Implement P2P connection logic here
+        pass
